core: allow for implicit setup of the cache app
Calling `ctx.App()` will implicitly load and provision the cache app if not already done. The `Provision` method is added to instantiate the map to avoid encountering nil maps. This change makes the explicit mention of `ja3` as a global option unnecessary. The cache is creatad implicitly when the listener or the handler are configured.